+++ src/components/Header/Header.stories.tsx
@@ -1,6 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
 import Header from "./Header";
+import Pagination from "../Pagination/Pagination";
 
 const meta: Meta<typeof Header> = {
   title: "Components/Temp/Header",
@@ -14,5 +15,11 @@ type Story = StoryObj<typeof Header>;
 export const WithControls: Story = {
   args: {},
   parameters: {},
-  render: () => <Header />,
+  render: () => (
+    <>
+      <Header />
+      <Pagination pageCount={10} />
+      <h1>test </h1>
+    </>
+  ),
 };